[ fix ] 출결 모달 수정
